compliance_agent/requirements.txt
customtkinter>=5.2.0
transformers>=4.30.0
torch>=2.0.0
openai-whisper>=20230314
PyMuPDF>=1.23.0
spacy>=3.6.0
nltk>=3.8.0
jieba>=0.42.1
pandas>=2.0.0
numpy>=1.24.0
